+++ api/api/app.py
@@ -131,6 +131,7 @@ def make_app(config=None, database_writable=False, **kw):
         InternalPuzzleDetailsByIdView,
     )
     from api.puzzle_list import (
+        InternalPuzzleRenderedResourcesListView,
         PuzzleListView,
         PlayerPuzzleListView,
         GalleryPuzzleListView,
@@ -314,6 +315,9 @@ def make_app(config=None, database_writable=False, **kw):
     # support a sqlite client/server model. For now, these are mostly for just
     # being able to write to the database since only the api has database
     # connection with write mode on.
+
+    app.add_url_rule("/internal/puzzle-rendered-resources-list/", view_func=InternalPuzzleRenderedResourcesListView.as_view("internal-puzzle-rendered-resources-list"))
+
     app.add_url_rule(
         "/internal/puzzle/<puzzle_id>/details/",
         view_func=InternalPuzzleDetailsView.as_view("internal-puzzle-details"),