+++ CHANGELOG.md
@@ -38,6 +38,7 @@ manage remote servers.
 - Terraform deployment for a development environment. Work in progress of [issue #92](https://github.com/jkenlooper/puzzle-massive/issues/92)
 - Support for referral content in footer.
 - Default vagrant setup uses a local cdn that is backed by a local fake s3 server.
+- Script to move puzzle resources to be from s3 or back to being locally hosted.
 
 ### Changed